warn if Requires-Python cannot be setup
move patching to a module
